+++ src/RuleSet/DeclarationBlock.php
@@ -60,8 +60,9 @@ class DeclarationBlock extends RuleSet
             $selectors = [];
             $selectorParts = [];
             $stringWrapperCharacter = null;
+            $functionNestingLevel = 0;
             $consumedNextCharacter = false;
-            static $stopCharacters = ['{', '}', '\'', '"', ','];
+            static $stopCharacters = ['{', '}', '\'', '"', '(', ')', ','];
             do {
                 if (!$consumedNextCharacter) {
                     $selectorParts[] = $oParserState->consume(1);
@@ -81,8 +82,21 @@ class DeclarationBlock extends RuleSet
                             }
                         }
                         break;
-                    case ',':
+                    case '(':
                         if (!\is_string($stringWrapperCharacter)) {
+                            ++$functionNestingLevel;
+                        }
+                        break;
+                    case ')':
+                        if (!\is_string($stringWrapperCharacter)) {
+                            if ($functionNestingLevel <= 0) {
+                                throw new UnexpectedTokenException('anything but', ')');
+                            }
+                            --$functionNestingLevel;
+                        }
+                        break;
+                    case ',':
+                        if (!\is_string($stringWrapperCharacter) && $functionNestingLevel === 0) {
                             $selectors[] = \implode('', $selectorParts);
                             $selectorParts = [];
                             $oParserState->consume(1);
@@ -91,6 +105,9 @@ class DeclarationBlock extends RuleSet
                         break;
                 }
             } while (!\in_array($nextCharacter, ['{', '}'], true) || \is_string($stringWrapperCharacter));
+            if ($functionNestingLevel !== 0) {
+                throw new UnexpectedTokenException(')', $nextCharacter);
+            }
             $selectors[] = \implode('', $selectorParts); // add final or only selector
             $oResult->setSelectors($selectors, $oList);
             if ($oParserState->comes('{')) {